requirements.txt
streamlit
numpy
panda
matplotlib
